fstr_pereval/requirements.txt
Django
psycopg2
Pillow
drf-writable-nested
DjangoRestFramework
PostgreSQL